do not install yt-dlp if its already present

+++ install/tools/yt-dlp.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-sudo wget -O /bin/yt-dlp https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp
-sudo chmod +x /bin/yt-dlp
+if [ ! -f '/bin/yt-dlp' ]; then
+    sudo wget -O /bin/yt-dlp https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp
+    sudo chmod +x /bin/yt-dlp
+fi
